Rename `showexceptions.rb` to `show_exceptions.rb`.
