+++ padrino-core/lib/padrino-core/application.rb
@@ -1,7 +1,7 @@
 require 'padrino-core/application/flash'
 require 'padrino-core/application/rendering'
 require 'padrino-core/application/routing'
-require 'padrino-core/application/showexceptions'
+require 'padrino-core/application/show_exceptions'
 require 'padrino-core/application/authenticity_token'
 
 module Padrino
