+++ padrino-core/lib/padrino-core/application/show_exceptions.rb
